Moved special purpose basis extenders to ring
